requirements.txt
anyio==3.5.0
asgiref==3.5.0
asyncio==3.4.3
certifi==2021.10.8
chardet==4.0.0
click==8.0.4
fastapi==0.75.0
fyers-apiv2==2.0.5
h11==0.13.0
idna==2.10
pydantic==1.9.0
requests==2.25.1
sniffio==1.2.0
starlette==0.17.1
tornado==6.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
websocket-client==1.2.1
websockets==8.1
